backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
rdkit-pypi>=2022.9.5
python-multipart
pydantic>=2.0.0
google-generativeai>=0.3.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy<2.0.0
